requirements.txt
feedparser
openai
yagmail
python-dotenv
